requirements.txt
flask==3.0.3
python-dotenv==1.0.1
langchain==0.1.12
langchain-openai==0.0.8
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0 